moving empty PostingsEnum creation into a static public method
